+++ examples/Blasphemos/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes
 
+## [18.05.2026] - v822
+- level2 and level3 were split into 2 levels each
+
+
+## [18.05.2026] - v822
+- level2 and level3 were split into 2 levels each
+
+
 ## [18.05.2026] - v821
 - I don't think I could fix anything